chore: Refactor client code, move to admin folder
Used vite build configuration to have 2 entry files (added client.html).
Reduced the repo complexity by having one less package.

Also we removed a lot of duplicated configurations files and only kept the actual source files
