--- packages/admin/.vscode/settings.json
@@ -1,6 +1,0 @@
-{
-  "editor.formatOnSave": false,
-	"editor.codeActionsOnSave": {
-		"source.fixAll.eslint": true
-	}
-}
