+++ packages/admin/package.json
@@ -2,10 +2,10 @@
   "name": "@goncy/twincy-admin",
   "version": "0.0.0",
   "scripts": {
-    "dev": "vite --port=6602",
+    "dev": "vite --port=6601",
     "build": "tsc && vite build",
     "prestart": "npm run build",
-    "start": "vite preview --port=6602",
+    "start": "vite preview --port=6601",
     "tauri": "tauri"
   },
   "dependencies": {
