--- packages/client/.editorconfig
@@ -1,9 +1,0 @@
-root = true
-
-[*]
-insert_final_newline = true
-charset = utf-8
-indent_style = space
-indent_size = 2
-trim_trailing_whitespace = true
-max_line_length = 80
