--- packages/client/.eslintignore
@@ -1,5 +1,0 @@
-node_modules/
-coverage/
-.idea/
-.vscode/
-build/
