--- packages/client/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@goncy/twincy-client",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite --port=6601",
-    "build": "tsc && vite build",
-    "prestart": "npm run build",
-    "start": "vite preview --port=6601"
-  },
-  "dependencies": {
-    "framer-motion": "^4.1.17",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "socket.io-client": "^4.1.2"
-  },
-  "devDependencies": {
-    "@types/node": "^14.17.2",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.14.1",
-    "@typescript-eslint/parser": "^4.14.1",
-    "@vitejs/plugin-react-refresh": "^1.1.0",
-    "eslint": "^7.19.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-cypress": "^2.11.2",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "prettier": "^2.2.1",
-    "sass": "^1.32.5",
-    "typescript": "^4.1.2",
-    "vite": "^2.0.5"
-  }
-}
