--- packages/client/src/App/index.ts
@@ -1,1 +1,0 @@
-export {default} from "./App";
